refactor: view에서 우승자 찾던 것 비즈니스로 보고 도메인으로 옮김
