+++ src/main/java/game/Cars.java
@@ -5,6 +5,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import static view.ResultView.WINNERS_DELIMITER;
+
 public class Cars {
     private final List<Car> cars;
 
@@ -29,5 +31,33 @@ public class Cars {
     public Map<String, Integer> positionByName() {
         return this.cars.stream()
                 .collect(Collectors.toMap(Car::name, Car::position));
+    }
+
+    public List<String> winners() {
+        int maxPosition = calculateMaxPosition();
+        return this.cars.stream()
+                .filter(it -> it.position() == maxPosition)
+                .map(Car::name)
+                .collect(Collectors.toUnmodifiableList());
+    }
+
+    private int calculateMaxPosition() {
+        return this.cars.stream()
+                .map(Car::position)
+                .max(Integer::compareTo)
+                .orElseThrow(() -> new IllegalArgumentException("우승자 위치를 찾을 수 없습니다."));
+    }
+
+    public static void printResult(Cars cars) {
+        Map<String, Integer> positionByName = cars.positionByName();
+        int maxPosition = positionByName.values().stream()
+                .max(Integer::compareTo)
+                .orElseThrow(IllegalArgumentException::new);
+        String winners = positionByName.entrySet()
+                .stream()
+                .filter(entry -> entry.getValue() == maxPosition)
+                .map(Map.Entry::getKey)
+                .collect(Collectors.joining(WINNERS_DELIMITER));
+        System.out.println(winners + "가 최종 우승했습니다.");
     }
 }
